tests/test_main.py
import unittest
from bhaskar_agent.app.main import some_function  # Replace with actual function to test

class TestMain(unittest.TestCase):

    def test_some_function(self):
        self.assertEqual(some_function(), expected_result)  # Replace with actual test case

if __name__ == '__main__':
    unittest.main()